Create a separate module for interoperable functions with gammapy objects and add frame variable to SpatialModelConfig to read from the config
