+++ asgardpy/data/target.py
@@ -7,8 +7,6 @@ from typing import List
 
 import astropy.units as u
 import numpy as np
-from astropy.coordinates import SkyCoord
-from gammapy.maps import Map
 from gammapy.modeling import Parameter, Parameters
 from gammapy.modeling.models import (
     SPATIAL_MODEL_REGISTRY,
@@ -22,8 +20,12 @@ from gammapy.modeling.models import (
     create_fermi_isotropic_diffuse_model,
 )
 
-from asgardpy.base.base import AngleType, BaseConfig, PathType
+from asgardpy.base.base import AngleType, BaseConfig, FrameEnum, PathType
 from asgardpy.base.geom import SkyPositionConfig
+from asgardpy.gammapy.interoperate_models import (
+    xml_spatial_model_to_gammapy,
+    xml_spectral_model_to_gammapy,
+)
 
 __all__ = [
     "BrokenPowerLaw2SpectralModel",
@@ -38,12 +40,8 @@ __all__ = [
     "create_gal_diffuse_skymodel",
     "create_iso_diffuse_skymodel",
     "create_source_skymodel",
-    "params_renaming_to_gammapy",
-    "params_rescale_to_gammapy",
     "read_models_from_asgardpy_config",
     "set_models",
-    "xml_spatial_model_to_gammapy",
-    "xml_spectral_model_to_gammapy_params",
 ]
 
 
@@ -88,6 +86,7 @@ class SpatialModelConfig(BaseConfig):
     """
 
     type: str = ""
+    frame: FrameEnum = FrameEnum.icrs
     parameters: List[ModelParams] = [ModelParams()]
 
 
@@ -431,6 +430,8 @@ def read_models_from_asgardpy_config(config):
     spatial_model: `gammapy.modeling.models.SpatialModel`
         Spatial Model components of a gammapy SkyModel object.
     """
+    # SkyModel is the first component
+    # FoVBackgroundModel is the second component
     model_config = config.components[0]
 
     # Spectral Model
@@ -511,313 +512,14 @@ def config_to_dict(model_config):
         par_dict["frozen"] = par.frozen
         model_dict["parameters"].append(par_dict)
 
+    # For spatial model, include frame info
+    try:
+        frame_ = getattr(model_dict, "frame")
+        model_dict["frame"] = frame_
+    except:
+        pass
+
     return model_dict
-
-
-def params_renaming_to_gammapy(
-    params_1_name, params_gpy, spectrum_type, params_1_base_model="Fermi-XML"
-):
-    """
-    Reading from a given parameter name, get basic parameters details like name,
-    unit and is_norm as per Gammapy definition.
-
-    This function collects all such switch cases, based on the base model of the
-    given set of parameters.
-    """
-    if params_1_base_model == "Fermi-XML":
-        if params_1_name in ["norm", "prefactor", "integral"]:
-            params_gpy["name"] = "amplitude"
-            params_gpy["unit"] = "cm-2 s-1 TeV-1"
-            params_gpy["is_norm"] = True
-
-        if params_1_name in ["scale", "eb"]:
-            params_gpy["name"] = "reference"
-
-        if params_1_name in ["breakvalue"]:
-            params_gpy["name"] = "ebreak"
-
-        if params_1_name in ["lowerlimit"]:
-            params_gpy["name"] = "emin"
-
-        if params_1_name in ["upperlimit"]:
-            params_gpy["name"] = "emax"
-
-        if params_1_name in ["cutoff", "expfactor"]:
-            params_gpy["name"] = "lambda_"
-            params_gpy["unit"] = "TeV-1"
-
-        if params_1_name in ["index"]:
-            params_gpy["name"] = "index"
-
-        if params_1_name in ["index1"]:
-            if spectrum_type in ["PLSuperExpCutoff", "PLSuperExpCutoff2"]:
-                params_gpy["name"] = "index"
-            else:
-                params_gpy["name"] = "index1"  # For spectrum_type == "BrokenPowerLaw"
-
-        if params_1_name in ["indexs"]:
-            params_gpy["name"] = "index_1"  # For spectrum_type == "PLSuperExpCutoff4"
-
-        if params_1_name in ["index2"]:
-            if spectrum_type == "PLSuperExpCutoff4":
-                params_gpy["name"] = "index_2"
-            elif spectrum_type in ["PLSuperExpCutoff", "PLSuperExpCutoff2"]:
-                params_gpy["name"] = "alpha"
-            else:
-                params_gpy["name"] = "index2"  # For spectrum_type == "BrokenPowerLaw"
-
-        if params_1_name in ["expfactors"]:
-            params_gpy["name"] = "expfactor"
-
-    return params_gpy
-
-
-def params_rescale_to_gammapy(params_gpy, spectrum_type, en_scale_1_to_gpy=1.0e-6, keep_sign=False):
-    """
-    Rescaling parameters to be used with Gammapy standard units, by using the
-    various physical factors (energy for now), compared with the initial set of
-    parameters as compared with Gammapy standard units.
-
-    Also, scales the value, min and max of the given parameters, depending on
-    their Parameter names.
-    """
-    if params_gpy["name"] in ["reference", "ebreak", "emin", "emax"]:
-        params_gpy["unit"] = "TeV"
-        params_gpy["value"] = (
-            float(params_gpy["value"]) * float(params_gpy["scale"]) * en_scale_1_to_gpy
-        )
-        if "error" in params_gpy:
-            params_gpy["error"] = (
-                float(params_gpy["error"]) * float(params_gpy["scale"]) * en_scale_1_to_gpy
-            )
-        params_gpy["min"] = (
-            float(params_gpy["min"]) * float(params_gpy["scale"]) * en_scale_1_to_gpy
-        )
-        params_gpy["max"] = (
-            float(params_gpy["max"]) * float(params_gpy["scale"]) * en_scale_1_to_gpy
-        )
-        params_gpy["scale"] = 1.0
-
-    if params_gpy["name"] in ["amplitude"]:
-        params_gpy["value"] = (
-            float(params_gpy["value"]) * float(params_gpy["scale"]) / en_scale_1_to_gpy
-        )
-        if "error" in params_gpy:
-            params_gpy["error"] = (
-                float(params_gpy["error"]) * float(params_gpy["scale"]) / en_scale_1_to_gpy
-            )
-        params_gpy["min"] = (
-            float(params_gpy["min"]) * float(params_gpy["scale"]) / en_scale_1_to_gpy
-        )
-        params_gpy["max"] = (
-            float(params_gpy["max"]) * float(params_gpy["scale"]) / en_scale_1_to_gpy
-        )
-        params_gpy["scale"] = 1.0
-
-    if params_gpy["name"] in ["index", "index_1", "index_2", "beta"] and not keep_sign:
-        # Other than EBL Attenuated Power Law?
-        # spectral indices in gammapy are always taken as positive values.
-        val_ = float(params_gpy["value"]) * float(params_gpy["scale"])
-        if val_ < 0:
-            params_gpy["value"] = -1 * val_
-
-            # Reverse the limits while changing the sign
-            min_ = -1 * float(params_gpy["min"]) * float(params_gpy["scale"])
-            max_ = -1 * float(params_gpy["max"]) * float(params_gpy["scale"])
-            params_gpy["min"] = min(min_, max_)
-            params_gpy["max"] = max(min_, max_)
-            params_gpy["scale"] = 1.0
-
-    if params_gpy["name"] in ["lambda_"] and spectrum_type == "PLSuperExpCutoff":
-        # Original parameter is inverse of what gammapy uses
-        val_ = float(params_gpy["value"]) * float(params_gpy["scale"])
-        params_gpy["value"] = en_scale_1_to_gpy / val_
-        if "error" in params_gpy:
-            params_gpy["error"] = en_scale_1_to_gpy * float(params_gpy["error"]) / (val_**2)
-        min_ = en_scale_1_to_gpy / (float(params_gpy["min"]) * float(params_gpy["scale"]))
-        max_ = en_scale_1_to_gpy / (float(params_gpy["max"]) * float(params_gpy["scale"]))
-        params_gpy["min"] = min(min_, max_)
-        params_gpy["max"] = max(min_, max_)
-        params_gpy["scale"] = 1.0
-
-    if float(params_gpy["scale"]) != 1.0:
-        # Without any other modifications, but using the scale value
-        params_gpy["value"] = float(params_gpy["value"]) * float(params_gpy["scale"])
-        if "error" in params_gpy:
-            params_gpy["error"] = float(params_gpy["error"]) * float(params_gpy["scale"])
-        params_gpy["min"] = float(params_gpy["min"]) * float(params_gpy["scale"])
-        params_gpy["max"] = float(params_gpy["max"]) * float(params_gpy["scale"])
-        params_gpy["scale"] = 1.0
-
-    return params_gpy
-
-
-def xml_spectral_model_to_gammapy_params(params, spectrum_type, is_target=False, keep_sign=False):
-    """
-    Convert the Spectral Models information from XML model of FermiTools to Gammapy
-    standards and return Parameters list.
-    Details of the XML model can be seen at
-    https://fermi.gsfc.nasa.gov/ssc/data/analysis/scitools/source_models.html
-    and with examples at
-    https://fermi.gsfc.nasa.gov/ssc/data/analysis/scitools/xml_model_defs.html
-
-    Models from the XML model, not read are -
-    ExpCutoff, BPLExpCutoff, PLSuperExpCutoff3, Gaussian, BandFunction
-
-    Parameters
-    ----------
-    params: `gammapy.modeling.Parameters`
-        List of gammapy Parameter object of a particular Model
-    spectrum_type: str
-        Spectrum type as defined in XML. To be used only for special cases like
-        PLSuperExpCutoff, PLSuperExpCutoff2 and PLSuperExpCutoff4
-    is_target: bool
-        Boolean to check if the given list of Parameters belong to the target
-        source model or not.
-    keep_sign: bool
-        Boolean to keep the same sign on the parameter values or not.
-
-    Returns
-    -------
-    params_final: `gammapy.modeling.Parameters`
-        Final list of gammapy Parameter object
-    """
-    new_params = []
-
-    for par in params:
-        new_par = {}
-
-        for key_ in par.keys():
-            # Getting the "@par_name" for each parameter without the "@"
-            if key_ != "@free":
-                new_par[key_[1:].lower()] = par[key_]
-            else:
-                if par["@name"].lower() not in ["scale", "eb"]:
-                    new_par["frozen"] = (par[key_] == "0") and not is_target
-                else:
-                    # Never change frozen status of Reference Energy
-                    new_par["frozen"] = par[key_] == "0"
-            new_par["unit"] = ""
-            new_par["is_norm"] = False
-
-            # Using the nomenclature as used in Gammapy
-            new_par = params_renaming_to_gammapy(
-                par["@name"].lower(), new_par, spectrum_type, params_1_base_model="Fermi-XML"
-            )
-
-        # Some modifications on scaling/sign:
-        new_par = params_rescale_to_gammapy(
-            new_par, spectrum_type, en_scale_1_to_gpy=1.0e-6, keep_sign=keep_sign
-        )
-
-        if new_par["name"] == "alpha" and spectrum_type in [
-            "PLSuperExpCutoff",
-            "PLSuperExpCutoff2",
-        ]:
-            new_par["frozen"] = par["@free"] == "0"
-
-        # Read into Gammapy Parameter object
-        new_param = Parameter(name=new_par["name"], value=new_par["value"])
-        if "error" in new_par:
-            new_param.error = new_par["error"]
-        new_param.min = new_par["min"]
-        new_param.max = new_par["max"]
-        new_param.unit = new_par["unit"]
-        new_param.frozen = new_par["frozen"]
-        new_param._is_norm = new_par["is_norm"]
-
-        new_params.append(new_param)
-
-    params_final2 = Parameters(new_params)
-
-    # Modifications when different parameters are interconnected
-    if spectrum_type == "PLSuperExpCutoff2":
-        alpha_inv = 1 / params_final2["alpha"].value
-        min_sign = 1
-        if params_final2["lambda_"].min < 0:
-            min_sign = -1
-
-        params_final2["lambda_"].value = params_final2["lambda_"].value ** alpha_inv * 1.0e6
-        params_final2["lambda_"].min = min_sign * (
-            abs(params_final2["lambda_"].min) ** alpha_inv * 1.0e6
-        )
-        params_final2["lambda_"].max = params_final2["lambda_"].max ** alpha_inv * 1.0e6
-
-    return params_final2
-
-
-def xml_spatial_model_to_gammapy(aux_path, xml_spatial_model):
-    """
-    Read the spatial model component of the XMl model to Gammapy SpatialModel
-    object.
-
-    Details of the XML model can be seen at
-    https://fermi.gsfc.nasa.gov/ssc/data/analysis/scitools/source_models.html
-    and with examples at
-    https://fermi.gsfc.nasa.gov/ssc/data/analysis/scitools/xml_model_defs.html
-
-    Parameters
-    ----------
-    aux_path: `Path`
-        Path to the template diffuse models
-    xml_spatial_model: `dict`
-        Spatial Model component of a particular source from the XML file
-
-    Returns
-    -------
-    spatial_model: `gammapy.modeling.models.SpatialModel`
-        Gammapy Spatial Model object
-    """
-    spatial_pars = xml_spatial_model["parameter"]
-
-    if xml_spatial_model["@type"] == "SkyDirFunction":
-        for par_ in spatial_pars:
-            if par_["@name"] == "RA":
-                lon_0 = f"{par_['@value']} deg"
-            if par_["@name"] == "DEC":
-                lat_0 = f"{par_['@value']} deg"
-        fk5_frame = SkyCoord(
-            lon_0,
-            lat_0,
-            frame="fk5",
-        )
-        gal_frame = fk5_frame.transform_to("galactic")
-        spatial_model = SPATIAL_MODEL_REGISTRY.get_cls("PointSpatialModel")().from_position(
-            gal_frame
-        )
-
-    elif xml_spatial_model["@type"] == "SpatialMap":
-        file_name = xml_spatial_model["@file"].split("/")[-1]
-        file_path = aux_path / f"Templates/{file_name}"
-
-        spatial_map = Map.read(file_path)
-        spatial_map = spatial_map.copy(unit="sr^-1")
-
-        spatial_model = SPATIAL_MODEL_REGISTRY.get_cls("TemplateSpatialModel")(
-            spatial_map, filename=file_path
-        )
-
-    elif xml_spatial_model["@type"] == "RadialGaussian":
-        for par_ in spatial_pars:
-            if par_["@name"] == "RA":
-                lon_0 = f"{par_['@value']} deg"
-            if par_["@name"] == "DEC":
-                lat_0 = f"{par_['@value']} deg"
-            if par_["@name"] == "Sigma":
-                sigma = f"{par_['@value']} deg"
-
-        fk5_frame = SkyCoord(
-            lon_0,
-            lat_0,
-            frame="fk5",
-        )
-        gal_frame = fk5_frame.transform_to("galactic")
-
-        spatial_model = SPATIAL_MODEL_REGISTRY.get_cls("GaussianSpatialModel")(
-            lon_0=gal_frame.l, lat_0=gal_frame.b, sigma=sigma, frame="galactic"
-        )
-
-    return spatial_model
 
 
 def create_source_skymodel(config_target, source, aux_path):
@@ -889,7 +591,7 @@ def create_source_skymodel(config_target, source, aux_path):
                         )()
 
         # Read the parameter values from XML file to create SpectralModel
-        params_list = xml_spectral_model_to_gammapy_params(
+        params_list = xml_spectral_model_to_gammapy(
             spectrum,
             spectrum_type,
             is_target=is_source_target,
